Move footnotes into the table as inline notes
There's a limitation in Markdown that footnotes such as `[^6]` cannot
be rendered when found within table cells. Perplexity AI tells me that
this is not possible, and gives a few possible reasons:
https://www.perplexity.ai/search/why-can-t-i-use-markdown-footn-L.TQdIMSSmeCn4QugvC8tw#0

Rather than have these completely broken, I figured we should render
them properly. This is what I found which I considered to be a
reasonable solution.

In the process I also figured I should fix the references for the first
three footnotes on this page at the same time.